refactor: document code
