+++ src/main/kotlin/gg/ingot/iron/sql/params/SqlParamsBuilder.kt
@@ -1,0 +1,54 @@
+package gg.ingot.iron.sql.params
+
+import gg.ingot.iron.representation.ExplodingModel
+import gg.ingot.iron.transformer.ModelTransformer
+
+/**
+ * A builder for SQL parameters that allows for easy parameter creation and addition.
+ */
+class SqlParamsBuilder internal constructor(
+    val values: SqlParams
+) {
+    /** The models to add to the parameters. */
+    private val models = mutableListOf<ExplodingModel>()
+
+    /**
+     * Adds a model to the parameters.
+     * @param model The model to add to the parameters.
+     * @return The SqlParams instance for chaining.
+     */
+    operator fun plus(model: ExplodingModel): SqlParamsBuilder {
+        models.add(model)
+        return this
+    }
+
+    /**
+     * Builds the SQL parameters using the provided transformer.
+     * @param transformer The transformer to use to build the parameters.
+     * @return The built SQL parameters.
+     */
+    internal fun build(transformer: ModelTransformer): SqlParams {
+        val variables = values.toMutableMap()
+
+        for (model in models) {
+            val entity = transformer.transform(model::class)
+            for (field in entity.fields) {
+                variables[field.columnName] = transformer.getModelValue(model, field)
+            }
+        }
+
+        return variables
+    }
+}
+
+/**
+ * Creates a new SqlParams instance with the provided values.
+ * @param values The values to add to the parameters.
+ * @return The SqlParams instance for chaining.
+ */
+fun sqlParams(vararg values: Pair<String, Any?>): SqlParamsBuilder {
+    return SqlParamsBuilder(values.toMap().toMutableMap())
+}
+
+/** The parameters to use in a SQL query. */
+typealias SqlParams = MutableMap<String, Any?>
